Break down test statistics page in to component and page for reusability
