+++ src/Pixel.Automation.Web.Portal/Pages/Tests/StatisticsPage.razor
@@ -1,0 +1,14 @@
+﻿@page "/test-statistics/{TestId?}"
+
+<MudGrid Elevation="4" Justify="Justify.Center">
+    <MudItem xs="12">
+        <MudPaper Elevation="3" Class="d-flex flex-row pt-4 pl-3" Style="height:60px;">
+            <MudText Typo="Typo.h5" Color="Color.Primary">Test Statistics</MudText>           
+        </MudPaper>
+    </MudItem>
+</MudGrid>
+<br />
+@if(!string.IsNullOrEmpty(TestId))
+{
+    <Pixel.Automation.Web.Portal.Components.Tests.TestStatistics TestId="@TestId"/>
+}
